+++ src/context/AppDataContext.tsx
@@ -14,6 +14,8 @@ export interface AppSettings {
   localHlsEnabled?: boolean;
   languageSelected?: string;
   trackingEnabled?: boolean;
+  playbackRate?: number;
+  uiScale?: number;
 }
 
 interface AppDataContext {
@@ -23,7 +25,7 @@ interface AppDataContext {
 
 // @ts-ignore
 const defaultContext: AppDataContext = {
-  appSettings: {},
+  appSettings: {playbackRate: 1, uiScale: 1},
   updateSettings: () => {},
 };
 
@@ -50,6 +52,8 @@ function setSettings(settings: Partial<AppSettings>) {
 
   const curSettings = getSettings();
   const newValue: AppSettings = {
+    playbackRate: 1,
+    uiScale: 1,
     ...curSettings,
     ...settings,
   };
@@ -62,7 +66,7 @@ interface Props {
 
 export default function AppDataContextProvider({children}: Props) {
   const [settings, setSettingState] = useState<AppSettings>(
-    getSettings() ?? {},
+    getSettings() ?? {playbackRate: 1, uiScale: 1},
   );
 
   const updateSettings = useCallback((data: Partial<AppSettings>) => {