+++ src/navigation/SettingsNavigator.tsx
@@ -3,6 +3,8 @@ import {createNativeStackNavigator} from "@react-navigation/native-stack";
 import LanguageSelectorScreen from "../components/settings/screens/LanguageSelector";
 import PlayerResolutionSelectorScreen from "../components/settings/screens/PlayerResolutionSelector";
 import PlayerTypeSelectorScreen from "../components/settings/screens/PlayerSelector";
+import PlaybackSpeedSelectorScreen from "../components/settings/screens/PlaybackSpeedSelector";
+import UiScaleSelectorScreen from "../components/settings/screens/UiScaleSelector";
 import SettingsScreen from "../screens/SettingsScreen";
 
 import TrackingSelector from "@/components/settings/screens/TrackingSelector";
@@ -13,6 +15,8 @@ export type SettingsStackParamList = {
   PlayerSelector: undefined;
   PlayerResolutionSelector: undefined;
   TrackingSelector: undefined;
+  PlaybackSpeedSelector: undefined;
+  UiScaleSelector: undefined;
 };
 
 const Stack = createNativeStackNavigator<SettingsStackParamList>();
@@ -34,6 +38,11 @@ export default function SettingsNavigator() {
         component={PlayerResolutionSelectorScreen}
       />
       <Stack.Screen name={"TrackingSelector"} component={TrackingSelector} />
+      <Stack.Screen
+        name={"PlaybackSpeedSelector"}
+        component={PlaybackSpeedSelectorScreen}
+      />
+      <Stack.Screen name={"UiScaleSelector"} component={UiScaleSelectorScreen} />
     </Stack.Navigator>
   );
 }